Fixed problem with key generation

+++ src/main/java/net/metricspace/crypto/ciphers/stream/salsa/Salsa20KeyGeneratorSpi.java
@@ -50,7 +50,7 @@ public final class Salsa20KeyGeneratorSpi extends SalsaFamilyKeyGeneratorSpi {
      * {@inheritDoc}
      */
     @Override
-    protected final SecretKey engineGenerateKey(final int[] data) {
+    protected final SecretKey engineGenerateKey(final byte[] data) {
         return new Salsa20CipherSpi.Key(data);
     }
 }
